core: use textarea
